Item Tiers 0-B: funnel item attribute field access through accessors (#38)
Prefactor per PLANS/ItemTiers_Implementation_Plan.md Cycle 0-B so the
Phase-1 struct reshape only touches accessor bodies. Zero behavior change
(grep-verified: no m_instance.<attr> access outside the two shared headers).

- ItemInstanceData.h: canonical attribute funnel — raw getters/setters
  (get_prefix_type, set_prefix, set_enchant_bonus, is_custom_made,
  has_prefix, ...) + copy_attributes_to/load_attributes_from templates
  moved down from CItem
- Item.h: CItem keeps enum-typed legacy views; all attribute helpers
  now delegate to the instance-data funnel
- Server: ItemManager (~70 sites), Game.cpp, AccountSqliteStore binds,
  CraftingManager, TradingPostStore
- Client: packet->instance copy blocks collapse to load_attributes_from
  (NetworkMessages_Items x4, Game.cpp, DialogBox_TradingPost); reads
  funneled in ItemNameFormatter, InventoryManager, dialogs, Screen_OnGame*
- Wire/DB row-struct field names intentionally untouched (Phase 1-C/1-E)
- Drive-by: fix pre-existing Linux/GCC break in DialogBox_NpcSpawner
  (packed field vs std::format reference binding)
- Adds the implementation plan doc + CHANGELOG entries for 0-A and 0-B

Verified: Windows -Target All, Linux server, Linux client all green.
Part of epic #36; unblocks #40.

+++ Sources/Client/DialogBox_NpcSpawner.cpp
@@ -137,7 +137,7 @@ void DialogBox_NpcSpawner::on_draw()
 
 		auto color = (idx == m_selected_index) ? GameColors::UIPaleYellow
 			: (hover ? GameColors::UIWhite : GameColors::UIMagicBlue);
-		auto label = std::format("[{}] {}", entry.npc_id, entry.name);
+		auto label = std::format("[{}] {}", static_cast<int>(entry.npc_id), entry.name);
 		hb::shared::text::draw_text_aligned(GameFont::Default,
 			sX + layout::content_x1 + 6, ry, layout::content_w - 12, 15,
 			label.c_str(),
